requirements.txt
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
google-genai>=1.0,<2
python-dotenv>=1.0,<2
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
streamlit>=1.37,<2
joblib>=1.4,<2
scikit-learn>=1.5,<2
xgboost>=2.1,<3
numpy>=1.26,<3
pandas>=2.2,<3
